% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sdo.R
\name{parse_sdo}
\alias{parse_sdo}
\title{Extract information about DataDownload objects from schema.org-based prov}
\usage{
parse_sdo(json)
}
\arguments{
\item{json}{input JSON-LD file}
}
\value{
A data.frame with core metadata about each DataDownload object
}
\description{
Extract information about DataDownload objects from schema.org-based prov
}
