% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{prov}
\alias{prov}
\title{generate provenance information}
\usage{
prov(
  data_in = NULL,
  code = NULL,
  data_out = NULL,
  meta = NULL,
  creator = NULL,
  title = NULL,
  description = NULL,
  issued = as.character(Sys.Date()),
  license = "https://creativecommons.org/publicdomain/zero/1.0/legalcode",
  schema = c("http://schema.org", "http://www.w3.org/ns/dcat"),
  embed_actions = FALSE,
  ...
)
}
\arguments{
\item{data_in}{path or URI for input data}

\item{code}{path or URI for code}

\item{data_out}{path or URI to output data}

\item{meta}{path or URI to metadata describing the workflow}

\item{creator}{URI, list node, or text for creator}

\item{title}{Dataset title, character string}

\item{description}{Dataset description, character string}

\item{issued}{publication date, as Date or character object}

\item{license}{URL to a copyright license}

\item{schema}{Use schema.org or DCAT2 schema? See details.}

\item{embed_actions}{should we incldue schema:Action to create?}

\item{...}{additional named elements passed to Dataset}
}
\description{
generate provenance information
}
