% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sdo.R
\name{to_sdo}
\alias{to_sdo}
\title{Convert DCAT2 Provenance to Schema.org}
\usage{
to_sdo(prov, con = stdout())
}
\arguments{
\item{prov}{JSON-LD file in DCAT2 format}

\item{con}{Output, a \link{connection} object or character string}
}
\description{
Convert DCAT2 Provenance to Schema.org
}
\examples{
prov <- system.file("examples", "prov.json", package="prov")
dcat_to_sdo(prov)

}
