% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_tsv.R
\name{write_prov_tsv}
\alias{write_prov_tsv}
\title{write provenance as tsv file}
\usage{
write_prov_tsv(
  data_in = NULL,
  code = NULL,
  data_out = NULL,
  meta = NULL,
  issued = as.character(Sys.Date()),
  theme = NA_character_,
  keyword = NA_character_,
  provdb = "prov.tsv",
  append = fs::file_exists(provdb)
)
}
\arguments{
\item{data_in}{path or URI for input data}

\item{code}{path or URI for code}

\item{data_out}{path or URI to output data}

\item{meta}{path or URI to metadata describing the workflow}

\item{issued}{publication date, as Date or character object}

\item{theme}{theme name}

\item{keyword}{keyword}

\item{provdb}{path to output JSON file, default "prov.json"}

\item{append}{Should we append to existing json or overwrite it?}
}
\description{
Write provenance metadata to a table. While this format is less expressive
than JSON-LD, it is also more efficient, particularly when data is appended.
At this time, only documents dcat:Distribution (schema:DataDownload) types.
Column names are taken from DCAT but no official namespace is set.
}
